+++ utils/test/testinput/gdas_rads2ioda.yaml
@@ -2,5 +2,6 @@ provider: RADS
 window begin: 2018-04-15T06:00:00Z
 window end: 2018-04-15T12:00:00Z
 output file: rads_adt_3b_2021181.ioda.nc
+#ocean basin: RECCAP2_region_masks_all_v20221025.nc
 input files:
 - rads_adt_3b_2021181.nc4
